Implement the save to database
move save to excel and save to db to another file
